代码更新

+++ src/com/ag777/util/db/model/interf/Column.java
@@ -9,5 +9,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Column {
 	String name();
-	long size() default -1;
+	int size() default -1;
 }
